requirements.txt
# Requirements per Planet AI - Analisi Immobiliare
# Streamlit Cloud compatible

# Core
streamlit>=1.42.0
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
python-dateutil

# Visualizzazione
plotly>=5.17.0
altair
matplotlib>=3.7.0
Pillow

# Data processing
protobuf

# Web scraping
beautifulsoup4>=4.12.0

# Geocoding
geopy>=2.4.0

# Geospatial (per OMI KML)
fastkml==0.12
pyproj
lxml
pygeoif

# Report generation
python-docx>=1.1.0

# AI Analysis (opzionale)
anthropic>=0.39.0

# Mappe interattive (NUOVO)
folium>=0.14.0
streamlit-folium>=0.15.0